requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
mysql-connector-python==9.1.0
pydantic==2.9.2
python-dotenv==1.0.1